Create lib/scraper.rb file, move Scraper class to it from bin/main.rb
